[의정] 찜, 테마 기능 수정 및 모델 수정

+++ controllers/LikeController.js
@@ -8,22 +8,30 @@ exports.Like = async function(req, res) {
         //요청 헤더 내의 authorization 헤더에서 토큰 추출
         let token = req.headers.authorization.split('Bearer ')[1];
         //토큰과 일치하는 user id 가져오기
-        let id = await USER.findOne({
+        let result = await USER.findOne({
             attributes: ['id'], 
             where : {token: token}
         })
+        let id = result.dataValues.id;
+        if(!id){
+            res.status(400).json({
+                error: "사용자미상"
+            });
+            return
+        }
         //해당 음식점을 찜했는지 확인
-        await LIKE.findOne({where : {id: id, r_code: r_code}})
+        await LIKE.findOne({where : {id: id, r_code: req.body.r_code}})
         .then((data) => {
             if(data) { // 반환 데이터가 있다면 취소요청이 들어온 것
                 LIKE.destroy({
-                    where: {id: id, r_code: r_code}
+                    where: {id: id, r_code: req.body.r_code}
                 })
-                res.status(204).json({
-                    message: "찜이 취소됐습니다."
+                res.status(200).json({
+                    message: '찜이 취소됐습니다.'
                 });
             }
             else { //반환 데이터가 없다면 새로이 찜을 하겠다는 것
+                console.log(req.body.r_code);
                 LIKE.create({
                     id : id, 
                     r_code : req.body.r_code
